# chore(pricing): improve calculate prices performances