--- a/.changeset/olive-horses-judge.md
+++ b/.changeset/olive-horses-judge.md
@@ -1,0 +1,5 @@
+---
+"@medusajs/pricing": patch
+---
+
+chore(pricing): improve performances