--- a/packages/modules/pricing/src/models/price-list-rule.ts
+++ b/packages/modules/pricing/src/models/price-list-rule.ts
@@ -15,6 +15,10 @@ const PriceListRule = model
       on: ["price_list_id"],
       where: "deleted_at IS NULL",
     },
+    {
+      on: ["attribute"],
+      where: "deleted_at IS NULL",
+    },
   ])
 
 export default PriceListRule