--- a/packages/modules/pricing/src/models/price-rule.ts
+++ b/packages/modules/pricing/src/models/price-rule.ts
@@ -18,6 +18,14 @@ const PriceRule = model
       where: "deleted_at IS NULL",
       unique: true,
     },
+    {
+      on: ["attribute", "value"],
+      where: "deleted_at IS NULL",
+    },
+    {
+      on: ["operator", "value"],
+      where: "deleted_at IS NULL",
+    },
   ])
 
 export default PriceRule