… packages/modules/pricing/src/migrations/.snapshot-medusa-pricing.json