**What**
Optimize the pricing retrieval query to improve performances (especially on large dataset)

It removed the overall sub query and many joins in favor of single join and very specific sub queries for rules check in order to benefit from indexes and pre filtered data.

Also, the produced query will be very much simpler if there is no context for rules check, in that case we only want prices that have 0 rules since all other prices would fail to match their expectations leading to more performant queries in these cases as well.

The resulting query plan is much cleaner and rely more on filtered data and indexes as well as sub plan that gets executed only when required and pre filtered the data


**TEST**
We ran some test on dataset containing about 1.2M prices and 1.2M price rules
1. The original query takes about 718ms execution time with a planning of 0.5ms
2. The new query takes about 17ms with a slightly higher execution plan time of 3.7ms

Overall this indicates a performance improvement at the query level of 42 times better performances form the original query